TEST: Add testing for each reducers files and Home and Details pages
